+++ src/processes/features/new-feature.starter.class.ts
@@ -2,7 +2,10 @@ import { FlowBase } from '@bohr/changelogger/libs/flow//flow-base.class';
 import { handleUncommittedChanges } from '@bohr/changelogger/processes/common-ops/handle-uncommitted-changes.function';
 import { FEATURE_NAME } from '@bohr/changelogger/questions/git/feature-name.constant';
 import { questionMaker } from '@bohr/changelogger/questions/question-maker.function';
-import { argv } from 'yargs';
+import { hideBin } from 'yargs/helpers';
+import yargs from 'yargs/yargs';
+
+const argv = yargs(hideBin(process.argv)).argv;
 
 export class NewFeatureStarter extends FlowBase {